moves allowed taxonomy to classmethod
