+++ peachjam/views/taxonomy.py
@@ -1,7 +1,6 @@
 from django.http import HttpResponseForbidden
 from django.shortcuts import Http404, get_object_or_404
 from django.views.generic import DetailView, TemplateView
-from guardian.shortcuts import get_objects_for_user
 
 from peachjam.models import Taxonomy
 from peachjam.views.generic_views import FilteredDocumentListView
@@ -13,53 +12,12 @@ class TaxonomyListView(TemplateView):
 
     def get(self, request, *args, **kwargs):
         context = self.get_context_data(**kwargs)
-        context["taxonomies"] = Taxonomy.dump_bulk()
+        context["taxonomies"] = Taxonomy.get_allowed_taxonomies(request.user)["tree"]
         context["taxonomy_url"] = "taxonomy_detail"
         return self.render_to_response(context)
 
 
-class TaxonomyFirstLevelView(DetailView):
-    template_name = "peachjam/taxonomy_first_level_detail.html"
-    model = Taxonomy
-    slug_url_kwarg = "topic"
-    context_object_name = "taxonomy"
-    navbar_link = "taxonomy"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["children"] = self.get_allowed_children()
-        context["taxonomy_link_prefix"] = "taxonomy"
-        return context
-
-    def get_allowed_children(self):
-        if self.request.user.is_authenticated:
-            allowed_taxonomies = set(
-                get_objects_for_user(
-                    self.request.user, "peachjam.view_taxonomy"
-                ).values_list("id", flat=True)
-            )
-        else:
-            allowed_taxonomies = []
-
-        children = self.object.get_children().values("id", "restricted")
-        exclude = []
-
-        for child in children:
-            is_restricted = child["restricted"]
-            is_allowed = child["id"] in allowed_taxonomies
-            if is_restricted and not is_allowed:
-                exclude.append(child["id"])
-
-        children = self.object.get_children().exclude(id__in=exclude)
-        return children
-
-
-class TaxonomyDetailView(FilteredDocumentListView):
-    template_name = "peachjam/taxonomy_detail.html"
-    navbar_link = "taxonomy"
-    # taxonomies may include legislation, so we want to show the latest expression only
-    latest_expression_only = True
-
+class AllowedTaxonomyMixin:
     def dispatch(self, request, *args, **kwargs):
         self.taxonomy = self.get_taxonomy()
         is_ancestor_restricted = self.taxonomy.get_ancestors().values_list(
@@ -68,8 +26,34 @@ class TaxonomyDetailView(FilteredDocumentListView):
         if self.taxonomy.restricted or any(is_ancestor_restricted):
             if not request.user.has_perm("peachjam.view_taxonomy", self.taxonomy):
                 return HttpResponseForbidden()
-        self.allowed_taxonomies = self.get_allowed_taxonomies(self.taxonomy)
+        self.allowed_taxonomies = Taxonomy.get_allowed_taxonomies(
+            request.user, root=self.taxonomy
+        )
         return super().dispatch(request, *args, **kwargs)
+
+
+class TaxonomyFirstLevelView(AllowedTaxonomyMixin, DetailView):
+    template_name = "peachjam/taxonomy_first_level_detail.html"
+    model = Taxonomy
+    slug_url_kwarg = "topic"
+    context_object_name = "taxonomy"
+    navbar_link = "taxonomy"
+
+    def get_taxonomy(self):
+        return self.get_object()
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["children"] = self.object.get_allowed_children(self.request.user)
+        context["taxonomy_link_prefix"] = "taxonomy"
+        return context
+
+
+class TaxonomyDetailView(AllowedTaxonomyMixin, FilteredDocumentListView):
+    template_name = "peachjam/taxonomy_detail.html"
+    navbar_link = "taxonomy"
+    # taxonomies may include legislation, so we want to show the latest expression only
+    latest_expression_only = True
 
     def get_taxonomy(self):
         root = get_object_or_404(Taxonomy, slug=self.kwargs["topic"])
@@ -85,50 +69,6 @@ class TaxonomyDetailView(FilteredDocumentListView):
         return (
             super().get_base_queryset().filter(taxonomies__topic__in=topics).distinct()
         )
-
-    def get_allowed_taxonomies(self, root):
-        if self.request.user.is_authenticated:
-            allowed_taxonomies = set(
-                get_objects_for_user(
-                    self.request.user, "peachjam.view_taxonomy"
-                ).values_list("id", flat=True)
-            )
-        else:
-            allowed_taxonomies = []
-
-        node_ids = []
-
-        def filter_nodes(node):
-            is_restricted = node.get("data", {}).get("restricted", False)
-            is_allowed = node.get("id") in allowed_taxonomies
-            if is_restricted and not is_allowed:
-                return None
-
-            node_ids.append(node["id"])
-
-            if "children" in node:
-                filtered_children = [
-                    child
-                    for child in (filter_nodes(child) for child in node["children"])
-                    if child is not None
-                ]
-                if filtered_children:
-                    node["children"] = filtered_children
-                else:
-                    node.pop("children", None)
-            return node
-
-        # Filter the entire tree
-        taxonomies = root.dump_bulk(root)
-        filtered_taxonomies = [
-            filtered_node
-            for filtered_node in (filter_nodes(node) for node in taxonomies)
-            if filtered_node is not None
-        ]
-        return {
-            "tree": filtered_taxonomies,
-            "pk_list": node_ids,
-        }
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
